requirements.txt
Django==1.8.4
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.0.2
autopep8==1.2
django-cors-headers==1.1.0
django-rest-auth==0.5.0
django-rest-swagger==0.3.4
djangorestframework==3.2.3
drf-extensions==0.2.7
pep8==1.6.2
python-dateutil==2.4.2
pytz==2015.4
requests==2.5.3
simplejson==3.8.0
six==1.9.0
sqlparse==0.1.16
tmdbsimple==1.3.0
